requirements.txt
flask
flask-mongoengine
pyyaml
uwsgidecorators
gunicorn
pymisp
pytz
python-dateutil